Added searchByName method to ServiceList
Added method to search list of services by sName and return CSV string of indexes of matching services. Discovered error in removeAllRows method of  GUI class where rows are not removed when there are fewer than 3 rows. At this stage this problem is not fixed but a slightly inefficient solution has been found.