All work done in BlueJ in College
Fixed the GUI such that it updates every minute when starting from a stop but not having a destination, this now appears to work. Added JourneyPlanning as was intended but this dose not seem to work yet. Changed some data structures. Added sorting and organisation classes. Commit before testing said classes with data.